+++ node/db-storage.d.ts
@@ -71,9 +71,9 @@ export declare class _NanoSQL_Storage {
     _clear(type: "all" | "hist", complete: Function): void;
     _delete(tableName: string, rowID: string | number, callBack?: (success: boolean) => void): void;
     _upsert(tableName: string, rowID: string | number | null, value: any, callBack?: (rowID: number | string) => void): void;
-    private _indexRead(tableName, rows, callBack);
+    private _indexRead(tableName, rows, callBack, getIndex?);
     _readRange(tableName: string, key: string, between: any[], callBack: (rows: DBRow[]) => void): void;
-    _read(tableName: string, row: string | number | Function, callBack: (rows: any[]) => void): void;
+    _read(tableName: string, row: string | number | Function, callBack: (rows: any[]) => void, readIndex?: boolean): void;
     _utility(type: "r" | "w", key: string, value?: any): any;
     _newTable(tableName: string, dataModels: DataModel[]): string;
 }