+++ node/index.js
@@ -265,7 +265,22 @@ var NanoSQLInstance = (function () {
             return this.backend._transaction("start");
     };
     NanoSQLInstance.prototype.endTransaction = function () {
+        var _this = this;
         this.doingTransaction = false;
+        Object.keys(this._models).forEach(function (table) {
+            if (table.indexOf("_") !== 0) {
+                _this.triggerEvent({
+                    table: table,
+                    query: [],
+                    time: new Date().getTime(),
+                    result: [],
+                    name: "change",
+                    actionOrView: "",
+                    changeType: "transaction",
+                    changedRows: []
+                }, ["change"]);
+            }
+        });
         if (this.backend._transaction)
             return this.backend._transaction("end");
     };