+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Read: ~100 records/ms
 - Add BTrees with `props:["trie"]` and `.extend("search",args)`.
 - Write LevelDB Cordova plugin.
 
+## [0.7.6] 4-8-2017
+- Secondary indexes and trie indexing now support multiple rows per entry. Existing secondary indexes should be rebuilt with `rebuildIndexes:true` before using this version.
+- Added naive event triggering on all tables after a transaction.
+
 ## [0.7.5] 4-7-2017
 - Removed transactions from index rebuilding, this was causing the index rebuild to fail on data sets in the thousands.
 - Removed automatic index sorting to speed up inserts again.